+++ api/use_cases/ocr/_createocrusecase.py
@@ -1,7 +1,11 @@
 from api.domain.model.entities import ModelType as RouterType
 from api.domain.ocr.entities import OCR, CreateOCRBody
 from api.domain.router.entities import Router
-from api.use_cases._forwarding import ForwardingCommand, ProviderRequestForwardingUseCase, ProviderRequestForwardingUseCaseSuccess
+from api.use_cases._providerrequestforwardingusecase import (
+    ForwardingCommand,
+    ProviderRequestForwardingUseCase,
+    ProviderRequestForwardingUseCaseSuccess,
+)
 from api.utils.variables import EndpointRoute
